requirements.txt
pbr>=0.6,!=0.7,<1.0
argparse
launchpadlib
requests>=1.1
PyYAML>=3.1.0
zuul
